+++ craft_parts/executor/part_handler.py
@@ -50,6 +50,7 @@ from craft_parts.utils.partition_utils import DEFAULT_PARTITION
 
 from . import filesets, migration
 from .environment import generate_step_environment
+from .errors import EnvironmentChangedError
 from .organize import organize_files
 from .step_handler import (
     StagePartitionContents,
@@ -102,6 +103,7 @@ class _Squasher:
         self.migrated_directories: dict[str | None, _MigratedContents] = {partition: {}}
         self._src_partition = partition
         self._default_partition = default_partition
+        self._distributed_paths: set[Path] = set()
         if filesystem_mount:
             self._filesystem_mount = filesystem_mount
 
@@ -123,8 +125,11 @@ class _Squasher:
             for entry in reversed(self._filesystem_mount):
                 # Only migrate content from the subdirectory indicated by the filesystem mounts
                 # entry
-                sub_path = entry.mount.lstrip("/")
+                sub_path = Path(entry.mount.lstrip("/"))
                 dst_partition = entry.device
+                logger.debug(
+                    "distribute content to %s, under %s", dst_partition, sub_path
+                )
 
                 # Migrate to the destination partition indicated by the filesystem mounts entry
                 self._migrate(
@@ -133,12 +138,15 @@ class _Squasher:
                     sub_path=sub_path,
                     dst_partition=dst_partition,
                 )
+                # If the sub path was really a sub path, record it.
+                if sub_path != Path("."):
+                    self._distributed_paths.add(sub_path)
         else:
             # Ignore the filesystem mounts and migrate from/to the same partition
             self._migrate(
                 srcdir=srcdir,
                 destdir=destdirs[self._src_partition],
-                sub_path="",
+                sub_path=Path(""),
                 dst_partition=self._src_partition,
             )
 
@@ -146,7 +154,7 @@ class _Squasher:
         self,
         srcdir: Path,
         destdir: Path,
-        sub_path: str,
+        sub_path: Path,
         dst_partition: str | None,
     ) -> None:
         """Actually migrate content from a source to a destination.
@@ -158,16 +166,14 @@ class _Squasher:
             srcdir / sub_path,
             destdir,
         )
-        logger.debug(f"excluding already migrated files: {self._all_migrated_files}")
-        visible_files = visible_files - self._all_migrated_files
-        logger.debug(
-            f"excluding already migrated dirs: {self._all_migrated_directories}"
-        )
-        visible_dirs = visible_dirs - self._all_migrated_directories
+
+        logger.debug("excluding content distributed to other partitions")
+        files = self._filter_already_distributed(visible_files)
+        dirs = self._filter_already_distributed(visible_dirs)
 
         layer_files, layer_dirs = migration.migrate_files(
-            files=visible_files,
-            dirs=visible_dirs,
+            files=files,
+            dirs=dirs,
             srcdir=srcdir / sub_path,
             destdir=destdir,
             oci_translation=True,
@@ -176,37 +182,30 @@ class _Squasher:
             self.migrated_files[dst_partition] = {}
 
         for f in layer_files:
-            src_path = str(Path(sub_path) / f)
+            src_path = str(sub_path / f)
             self.migrated_files[dst_partition][src_path] = f
 
         if dst_partition not in self.migrated_directories:
             self.migrated_directories[dst_partition] = {}
 
         for f in layer_dirs:
-            src_path = str(Path(sub_path) / f)
+            src_path = str(sub_path / f)
             self.migrated_directories[dst_partition][src_path] = f
 
-    @property
-    def _all_migrated_files(self) -> set[str]:
-        """Merge lists of files migrated to every partitions.
+    def _filter_already_distributed(self, visible_contents: set[str]) -> set[str]:
+        """Filter files in paths already distributed to other partitions."""
+        if not self._distributed_paths:
+            return visible_contents
+        contents: set[str] = set()
+        for content in visible_contents:
+            is_distributed = any(
+                migration.already_distributed(Path(content), distributed_path)
+                for distributed_path in self._distributed_paths
+            )
 
-        Return a list of paths relative to the source partition.
-        """
-        migrated_files: set[str] = set()
-        for m in self.migrated_files.values():
-            migrated_files |= set(m)
-        return migrated_files
-
-    @property
-    def _all_migrated_directories(self) -> set[str]:
-        """Merge lists of directories migrated to every partitions.
-
-        Return a list of paths relative to the source partition.
-        """
-        migrated_directories: set[str] = set()
-        for m in self.migrated_directories.values():
-            migrated_directories |= set(m)
-        return migrated_directories
+            if not is_distributed:
+                contents.add(content)
+        return contents
 
 
 class PartHandler:
@@ -588,6 +587,7 @@ class PartHandler:
         )
 
         self._migrate_overlay_files_to_prime()
+        default_partition = self._part_info.default_partition or DEFAULT_PARTITION
 
         if (
             self._part.spec.stage_packages
@@ -596,7 +596,7 @@ class PartHandler:
         ):
             prime_dirs = list(self._part.prime_dirs.values())
             primed_stage_packages = _get_primed_stage_packages(
-                contents.partitions_contents[DEFAULT_PARTITION].files,
+                contents.partitions_contents[default_partition].files,
                 prime_dirs=prime_dirs,
             )
         else:
@@ -607,8 +607,6 @@ class PartHandler:
             for p, c in contents.partitions_contents.items()
             if not self._part_info.is_default_partition(p)
         }
-
-        default_partition = self._part_info.default_partition or DEFAULT_PARTITION
 
         return states.PrimeState(
             partition=default_partition,
@@ -1146,6 +1144,39 @@ class PartHandler:
             )
             overlay_migration_state_path.unlink()
 
+    def _symlink_alias_to_default(self) -> None:
+        """Create directory and symlinks for the alias of the default partition.
+
+        These symlinks are never consumed by craft-parts. They are created to help
+        users debugging a build.
+        """
+        if not self._part_info.is_default_partition_aliased:
+            return
+        default_partition = self._part_info.default_partition
+        logger.debug("Create symlinks for %s", default_partition)
+        self._part_info.alias_partition_dir.mkdir(parents=True, exist_ok=True)
+
+        for src, dst in [
+            (self._part_info.parts_dir, self._part_info.parts_alias_symlink),
+            (self._part_info.stage_dir, self._part_info.stage_alias_symlink),
+            (self._part_info.prime_dir, self._part_info.prime_alias_symlink),
+            (self._part_info.overlay_dir, self._part_info.overlay_alias_symlink),
+        ]:
+            if dst.exists():
+                if not dst.is_symlink():
+                    # Between two runs of the lifecycle, the default partition alias name
+                    # can be changed to a previously concrete partition by the user.
+                    raise EnvironmentChangedError(
+                        f"cannot create symlinks {dst}, a concrete directory already exists."
+                    )
+                # The symlink already exists
+                continue
+            os.symlink(
+                src,
+                dst,
+                target_is_directory=True,
+            )
+
     def _make_dirs(self) -> None:
         dirs = [
             self._part.part_src_dir,
@@ -1158,9 +1189,12 @@ class PartHandler:
             self._part.part_run_dir,
             *self._part.stage_dirs.values(),
             *self._part.prime_dirs.values(),
+            *self._part.overlay_dirs.values(),
         ]
         for dir_name in dirs:
             os.makedirs(dir_name, exist_ok=True)
+
+        self._symlink_alias_to_default()
 
     def _fetch_stage_packages(self, *, step_info: StepInfo) -> list[str] | None:
         """Download stage packages to the part's package directory.