+++ tests/unit/executor/test_step_handler.py
@@ -37,6 +37,7 @@ from craft_parts.infos import (
 )
 from craft_parts.parts import Part
 from craft_parts.steps import Step
+from craft_parts.utils.partition_utils import DEFAULT_PARTITION
 
 from tests.unit.common_plugins import StrictTestPlugin
 
@@ -158,20 +159,26 @@ class TestStepHandlerBuiltins:
         deb = _get_host_architecture()
         triplet = _DEB_TO_TRIPLET[deb]
         if partitions is not None:
-            partition_script_lines = [
-                f'export CRAFT_DEFAULT_STAGE="{new_dir}/stage"',
-                f'export CRAFT_DEFAULT_PRIME="{new_dir}/prime"',
+            default_partition = partitions[0]
+            partition_script_lines = []
+            if default_partition != DEFAULT_PARTITION:
+                partition_script_lines = [
+                    f'export CRAFT_DEFAULT_STAGE="{new_dir}/stage"',
+                    f'export CRAFT_DEFAULT_PRIME="{new_dir}/prime"',
+                ]
+
+            partition_script_lines += [
+                f'export CRAFT_{default_partition.upper().translate({ord("-"): "_", ord("/"): "_"})}_STAGE="{new_dir}/stage"',
+                f'export CRAFT_{default_partition.upper().translate({ord("-"): "_", ord("/"): "_"})}_PRIME="{new_dir}/prime"',
                 *itertools.chain.from_iterable(
                     zip(
                         [
                             f'export CRAFT_{p.upper().translate({ord("-"): "_", ord("/"): "_"})}_STAGE="{new_dir}/partitions/{p}/stage"'
-                            for p in partitions
-                            if p != "default"
+                            for p in partitions[1:]
                         ],
                         [
                             f'export CRAFT_{p.upper().translate({ord("-"): "_", ord("/"): "_"})}_PRIME="{new_dir}/partitions/{p}/prime"'
-                            for p in partitions
-                            if p != "default"
+                            for p in partitions[1:]
                         ],
                     )
                 ),
@@ -255,16 +262,15 @@ class TestStepHandlerBuiltins:
         result = sh.run_builtin()
 
         step_contents = StepContents(stage=True)
-        step_contents.partitions_contents["default"] = StagePartitionContents(
+        default_partition = partitions[0] if partitions else "default"
+        step_contents.partitions_contents[default_partition] = StagePartitionContents(
             files={"subdir/bar", "foo"},
             dirs={"subdir"},
             backstage_files={"foo", "subdir/bar"},
             backstage_dirs={"subdir"},
         )
         if partitions:
-            for partition in partitions:
-                if partition == "default":
-                    continue
+            for partition in partitions[1:]:
                 step_contents.partitions_contents[partition] = StagePartitionContents(
                     files=set(), dirs=set()
                 )
@@ -289,13 +295,12 @@ class TestStepHandlerBuiltins:
         )
         result = sh.run_builtin()
         step_contents = StepContents()
-        step_contents.partitions_contents["default"] = StepPartitionContents(
+        default_partition = partitions[0] if partitions else "default"
+        step_contents.partitions_contents[default_partition] = StepPartitionContents(
             files={"subdir/bar", "foo"}, dirs={"subdir"}
         )
         if partitions:
-            for partition in partitions:
-                if partition == "default":
-                    continue
+            for partition in partitions[1:]:
                 step_contents.partitions_contents[partition] = StepPartitionContents(
                     files=set(), dirs=set()
                 )